requirements.txt
numpy
torch
scipy
pandas
scikit-learn
nibabel
tqdm
matplotlib
seaborn
Pillow
einops
